更新拦截器

+++ yn-core/src/main/java/com/yn/service/SolarPanelSerice.java
@@ -84,7 +84,7 @@ public class SolarPanelSerice {
 				
 				String typeName = type == 1 ? "单晶硅":"多晶硅";
 				
-				Integer conversionEfficiency =(Integer)object[3];
+				Double conversionEfficiency =(Double)object[3];
 
 					BigDecimal quality =(BigDecimal)object[4];
